Avoid leaking nodes in DAP test suites
Summary:
# Context
When running tests for the DAP server, we typically use `peer` to start debuggee nodes. We use `edb_test_support:start_peer_node()` and in `end_per_testcase()` we then call test `edb_test_support:stop_all_peer_nodes()` to do a clean-up

# Problem
Some testcases end with the debugee in a paused state. We've noticed that in that case, the debugee node is not actually stopped, but the node stays up after the test is finished.

# This diff
Because stopping edb_server resumes the paused processes, ensuring the server is stopped before attempting to stop the peer is enough to avoid the issue. We need to use `erpc:call()` instead of `peer:call()` since the latter blocks while the node is paused.

Reviewed By: thizanne

Differential Revision: D68015933

fbshipit-source-id: 2c350b2f71dc6d463deaa918f97aa8282b11ec74