[BE} Use more robust random node names in tests
Summary:
# Context
In tests, we start debugee nodes with random names.

# Problem
The random name is currently derived just from a timestamp. When running tests in parallel (e.g. stress-runs), there is a very low chance that node-names may collide.  While the chances in practice are very low, when debugging infrequent failures, this possibility can be distracting.

# This diff
So we use instead `peer:random_name/1`, which uses the OS pid of the process and a unique number per node for deriving the node name. A collision now would be so unlikely that we can simply disregard it from now on

Reviewed By: thizanne

Differential Revision: D68016286

fbshipit-source-id: c50716c3d59fdd17871b4022ef87e904e838ea87