Allow specifying a custom env to the launch/runInTerminal request
Summary:
The [`runInTerminal` request](https://microsoft.github.io/debug-adapter-protocol//specification.html#Reverse_Requests_RunInTerminal) allows the user to specify environment key-value pairs that are added to the default environment.

To be able to customize them, we define an equivalent structure in the *Launch Configuration*.

This will be useful, for example, to specify a custom Erlang installation path when launching the debuggee.

In principle (according to the specs), we should be able to "remove" values by selecting a "null" value, but that's out of scope for now.

Reviewed By: jcpetruzza

Differential Revision: D68016618

fbshipit-source-id: b828ec2f22051f631749d70abe2a3170073cec49