Specify required ERL_FLAGS during runInTerminal reverse request
Summary:
To be able to setup a functioning debugging session, we need to specify the `+D` flag during the `runInTerminal` request. We can do this directly on the server, to minimize the required configuration on the client side.

This approach can be refined later to be able to customize the key of the variable name and other options via a Launch configuration.

Reviewed By: jcpetruzza

Differential Revision: D68149788

fbshipit-source-id: b675c4cf1ed2607af5128f9c5a380dc219e10760