Test that we can see the stackframe (including top frame var) of a stepping process
Summary:
This checks in particular that we don't require to hit a user breakpoint in order to resolve variables (that used to be the case before the introduction of `edb_server_break:is_trapped`).

We add a dummy function that computes some arithmetic to avoid the current quirks of stepping in recursive functions.

Reviewed By: jcpetruzza

Differential Revision: D68508670

fbshipit-source-id: c3c51b3a918bb89f3741e79e96ea97e0cabd0f85