Do not register explicit hits for stepping breakpoints
Summary:
When processing a BP event, we now register an explicit hit only if we hit an explicit user breakpoint, not a stepping one.

Note that if a breakpoint exists for both reasons, "explicit takes priority", ie. we will generate an "user breakpoint hit" event and set the explicit hit.

We update tests accordingly. We change the ones that used to rely on get_breakpoints_hit to check current location to use stackframe information instead.

Reviewed By: jcpetruzza

Differential Revision: D68575378

fbshipit-source-id: 4fcce5b55d5a6365d781eb844841b80cc02ca27d