Import OTP toolchain wich changes in stack-frames format
Summary:
# Context
At the moment, the `erl_debugger:stack_frames/2` function gives us the mfa ./ line of each frame, plus the frame slots (Y-regs and exception-handlers).

# Problem
In order to implement stepping correctly, we need a way store the stack-frames at the moment the stepping started and, when hitting an internal breakpoint, compare it to the current stack-frames to understand if we are on the same function call, we are a frame above, below or we bounced.

However, with only mfa/line there is currently not enough information to know if we have "bounced" or not, as we could have more than once call to a function per line.

# This diff
So we extended the `erl_debugger:stack_frames()` so that it includes the "continuation pointer" of each stack-frame, essentially, the address of the code that will be executed when control goes back to that frame. THis wasy, we can uniquely distinguish frames, even if they correspond to code "on the same line".

Notice that here we just import the new toolchain, and make the code fixes so that the new way in which stack-frames is returned doesn't break existing code.

Changes on the OTP side are [here](https://github.com/jcpetruzza/otp/compare/edb-2025-01-23..edb-2025-01-24)

Reviewed By: thizanne

Differential Revision: D68637176

fbshipit-source-id: 1c7ee5202b742b8beec071cd7722fa13c9227020