Add edb:pause/0 function
Summary:
# Context
At the moment we can only pause the VM if a process hits a breakpoint, or a stepping process finishes stepping.

# Problem
It can be useful to be able to unconditionally "pause" the VM, e.g. to find out a non-responding node is doing

# This diff
Add a `edb:pause()` function that will just pause all processes being debugged.

Reviewed By: thizanne

Differential Revision: D68781255

fbshipit-source-id: 4184bb95f141a26c5f003be214c1c1f5f2746b6e