Use "paused" instead of "stopped" in events
Summary:
# Context
We are removing inconsistencies in the API of edb regarding "paused" vs "stopped". In particular, we want to use only "paused"  in edb's API (on the other side, the DAP protocol refers to both "stopped" and "pause", so there we can keep using "stopped" when it corresponds to the "stopped" event).

# This diff
We change it so that when the VM is paused, we now send a "paused" event, instead of a "stopped" one.

Reviewed By: thizanne

Differential Revision: D68781661

fbshipit-source-id: 65babf66963b38d234f983528be1530d0d80072c