Fix API inconsistency: not_paused vs not_stopped
Summary:
# Context
Some operations require the the VM to be paused (e.g. continue, step-over, etc), and when this is not the case, we need to reflect this in the return value

# Problem
We were being inconsistent and using "not_stopped" in some cases and "not_paused" in others.

# This diff
Use only `not_paused` for this. We seem to have this inconsistency in multiple places, where we use either "stopped" or "paused" for the same concept, we should probably move towards "paused" everywhere.

Reviewed By: acw224

Differential Revision: D68773224

fbshipit-source-id: 7180fff8a26b27d2aed17bc1298d6e47f168e9ad