Simplify "stopped" events
Summary:
# Context
When the VM is paused for whatever reason, we currently emit a "stopped" event. Which contains info such as which pid hit a bp or finished stepping.

# Problem
Actually, when the VM is paused, we generate two "stopped" events:

  - One for the pid that ht the BP, etc
  - Another to account for the processes that were suspended as part of the VM being paused

This was mostly for historical reasons, as the DAP server would initially generate two DAP events: one just for the pid in the BP, another one to pause all. This turned out to be incorrect and very inefficient, so we have since ignored the second "pause all" event. At this point, there is no good reason to keep that in the API

# This diff
We remove the "{paused, all}" and "{paused, Pid}" cases from the API and just let the "stopped" event imply "everything is paused". If in the future we want to have the two separate behaviours, we should extend the "stopped" event to explain where it is "global" or not.

NB. This is another step towards cleaning up the API in terms of "stopped" vs "paused"

Reviewed By: thizanne

Differential Revision: D68773223

fbshipit-source-id: 9f1bcc58b6676417bea990f030e8b4843f1c9185