Add a supervisor structure to edb_core
Summary:
# Context
Before splitting edb in two apps, the edb app would start the supervisor for the DAP server. So we couldn't have a supervisor structure for the "client" part (`edb`, `edb_node_monitor`).

# Problem
We were lazily starting `edb_node_monitor` on every `edb_node_monitor:call()`, which is not very idiomatic

# This diff
We add a `edb_core_sup` and ensure that it is started when the `edb_core` app is started.

Reviewed By: TheGeorge

Differential Revision: D68974473

fbshipit-source-id: b1dab2226cb4b55e9809af0ff9b33463bfa3ed7f