Native support for additional test fixtures for each group in edb_suite
Summary:
edb_suite groups of tests will automatically compile and load a `$GROUP.erl` file that will be used as the debuggee by each case in that group.

We augment this feature by also automatically loading any additional `$GROUP*.erl` file present in the same directory, the goal being to make it easier to write tests that imply different modules.

We do not have this kind of multi-module test yet so this should be a non-functional change.

Reviewed By: jcpetruzza

Differential Revision: D68727494

fbshipit-source-id: 692f0f1172b75319418e8a270ad799a6734cf4c1