Use explicit options when calling edb_test_support:start_peer_node/2
Summary:
# Context
We use this utility function to start debuggee nodes for tests. At the moment, we must specify either a "prefix" and a unique node name with that prefix is generated, or a exact node name to use. This is to cater for different scenarios

# Problem
There are two issues:

  1. In almost every case, we don't care about the node name, so we could just use a default.
  2. We want to add more options to control the creation of the peer node

# This diff
We introduce an explicit options argument for this function. In doing this, we make the node optional (so we can rely on the default when name is irrelevant)

Reviewed By: alanz

Differential Revision: D69193994

fbshipit-source-id: 462053f59f173772e9fc57e78f8217a57b941dce