Make passing the +D emu arg optional in edb_test_support:start_peer_node/2
Summary:
# Context
We use `edb_test_support:start_peer-node/2` in tests to start a debuggee node. Typically, the debuggee needs to be in debugging mode.

# Problem
If we want to test the behaviour when the node is not in debugging mode, we have no easy way to do it.

# This diff
Add an option to control whether we want to pass the option or not. It wil be used late in the stack

Reviewed By: alanz

Differential Revision: D69209755

fbshipit-source-id: 4f06b141bbd30fb428e6a2d7b11d8e60a847738d