Make copying code-path optional in edb_test_support:start_peer_node/2
Summary:
# Context
We use `edb_test_support:start_peer_node/2` to start debuggee nodes in tests. At the moment, the  `edb_core` app needs to be in the path of the debuggee for edb to work, so we extend the debuggee path with the test node's.

# Problem
We want to make it so that instead of relying on the debuggee having edb_core in the path, it is edb itself that injects the necessary code on attach. That means that in test-cases we will want to control whether edb_core is available on the debuggee or not.

# This diff
We introduce an option to control whether to copy the path when starting the node. For now we default it to `true` so that nothing breaks. Nothing changes the value at the moment.

Reviewed By: alanz

Differential Revision: D69196944

fbshipit-source-id: 20c86a184500fd47ddc498f7264ef884250a22a4