Debuggee node no longer need to have edb_core in path
Summary:
# Context
The debuggee node will run a `edb_server` process during the debugging session. The server is started, if needed, on attachment.

# Problem
Until now, it was the user's responsibility to ensure that the `edb_core` application was in the codepath of the debuggee, so that `edb_server` could be started. This is not very convenient.

# This diff
Now, `edb` will inject all the necessarily modules in the debuggee on attachment, making the setup more transparent.

Reviewed By: TheGeorge

Differential Revision: D69209756

fbshipit-source-id: e403516fe069f93b4ea255ac2d03ab4e9d62fd64