Add set_breakpoints(Module, Lines) command
Summary:
We add a set_breakpoints(Module, Lines) command to EDB that will atomically set all the breakpoints of the given module on the given lines (and clear the others).

The goal is to use it for the corresponding DAP request instead of a clear+add sequence that isn't atomic.

--
NB. I'm landing this on behalf of Thibault

Reviewed By: sopotc

Differential Revision: D69301612

fbshipit-source-id: 01ab5df25e7a43bb91f6b2cbd85cf64b0b912169

Co-authored-by: Thibault Suzanne <tsuzanne@meta.com>