Refactor edb_node_monitor to use gen_statem
Summary:
# Context
The edb_node_monitor is a gen_server that tracks the debuggee node. Conceptually, this server is a kind of state-machine, and we are planning to add yet more states.

# Problem
The logic is currently split accross various handle_* functions, that makes it harder to follow. We are also reimplementing functionality, like timeouts, that could be implemented on top of gen_statem primitives.

# This diff
We convert the module to implement the gen_statem behaviour. For now, we don't store anything useful in the "data" part of the state-machine, but we will move the subscription handling there next.

The conversion is mostly mechanic

Reviewed By: TheGeorge

Differential Revision: D69310794

fbshipit-source-id: 54a588126e33c8db154c507562258588c189dcfd