Turn the state of edb_node_monitor into a sum-type
Summary:
# Context
The edb_node_monitor is a gen_server that tracks the debuggee node. Conceptually, this server is a kind of state-machine, and we'd like to move it to implement `gen_statem` to make the logic more explicit and reuse functionality..

# Problem
Although not technically necessary, it is arguably more clear if the state of a state-machine is a sum-type, while the state of edb_node_monitor is currently a product of a sum (the possible "states") and a set of subscribers.

# This diff
We distribute the product over the sum, making the event-subscribers an argument of every state.

NB. After converting the module to use gen_statem, we'll pull-back the event-subscribers and put in the statem data. But let;s do one change at a time

Reviewed By: TheGeorge, thizanne

Differential Revision: D69317378

fbshipit-source-id: 5cab03b1f3d63234fc8fa102b203342b85b8ebb9