Use gen_statem timeout logic in edb_node_monitor
Summary:
# Context
We just refactored edb_node_monitor from being a gen_server to being a gen_statem; it's time to start using more of its functionality

# Problem
When we attaching to a node, we first go to the `attaching` state and from there we eventually transition to `attached`, once we are connected to the node, or eventually timeout, based on a user provided timeout.

At the moment we do all the timeout handling ourselves.

# This diff
We now just emit a `state_timeout` action when entering the `attaching` state. It is a no-op if the timeout is `infinity`, and it will get auto-cleared when we transition to another state.

Reviewed By: TheGeorge

Differential Revision: D69324610

fbshipit-source-id: 74a2e24f1b66aa5cc2c36dfaad138cee6ec026a2