Put node-monitor event-subscribers in state-machine data
Summary:
# Context
We just refactored edb_node_monitor from being a gen_server to being a gen_statem; it's time to start using more of its functionality

# Problem
Currently, in every state we hold a list of all event subscribers. Because it exists in every state, it makes more sense to put in the "data" part of the state-machine.

# This diff
We move the event-subscription to the "data" part of the state-machine, which simplifies a bit the code

Reviewed By: TheGeorge

Differential Revision: D69310793

fbshipit-source-id: 1bcb84126a6a54c2c8f464c349e7d5118c133e9f