Simplify type
Summary:
# Context
In D69324610 we made it so that instead of handling the timeout waiting for a node to be up, we rely on the gen_statem state_timeout mechanism.

# Problem
Previously we would expect to receive a `timeout` message, so this was part of the `info_message()` type. We no longer expect that message, so the summand is misleading

# This diff
Remove the redundant summand

Reviewed By: RobinMorisset

Differential Revision: D69592679

fbshipit-source-id: 0035167c55c1ea686efecf3fac8eb58550c5fd42