Allow for passing extra erl args when calling edb_test_support:start_peer_node/2
Summary:
# Context
We use `edb_test_support:start_peer_node/2` to spin up debuggee nodes in tests.

# Problem
As we add tests for reverse-attaching, we would like to be able to sometimes add things like `-eval blah` to the args of the peer (to inject the code to reverse attach), but atm there is no provision to do this

# This diff
We add a `extra_args` option, to specify this. It will be used later.

Reviewed By: RobinMorisset

Differential Revision: D69600440

fbshipit-source-id: 802321060fa691283fdd0c4b123b44b7cf6bf8db