Group testcases in edb_attachment_SUITE
Summary:
# Context
edb_attachment_SUITE contains all testcases related to attaching/detaching. At the moment they are not grouped.

# Problem
We'll be soon adding a "reverse-attach" functionality and we'll want testcases for it to end here as well. A flat structure makes it a bit harder to understand what functionality is being tested. etc/

# This diff
Split all testcases in "attaching" and "deteching" tests, so that we can later add "reverse-attaching" tests as well

Reviewed By: RobinMorisset

Differential Revision: D69599475

fbshipit-source-id: ac5f288412902b61ea82447c935dfc6beb0f2c03