Add option to specify connection cookie to edb:attach/1
Summary:
# Context
Attaching to a debuggee node means establishing a dist connection. Because the debuggee node is pre-existing, it may be using a different cookie than the one assigned to the debugger node. We should be able to specify what to use on connection.

# Problem
We currently offer the possibility to specify the cooke in the DAP request, but it is the DAP server who currently sets it. This should be handled by `edb:attach/1` itself.

# This diff
Add an optional `cookie` options to `edb:attach/1` and make the DAP server use that instead.

Reviewed By: TheGeorge

Differential Revision: D69856355

fbshipit-source-id: 31807e1f1215bdb1c483fd951f6efb8c34fef657