Fix issues in the generation of node names
Summary:
# Context
We create random node names in two places:

1.  in the context of the DAP server, before launching the debuggee, when we need to start the distribution stack; and
2. In tests, both for starting the distribution stack and creating peer nodes

# Problem
There were a couple of issues:

1. In the DAP server, we were always using `localhost` as host, which is problematic, as connections can then only be initiated from the debugger node (net_kernel:connect_node/1 fails with longames when the *target* is localhost)
2. We were also hardwiring "localhost" generating node names for tests, where we assume shortnames. This is still not a good practice, one should instead use `inet:gethostaddr/0` like `net_kernel:start()` does when no host is provided.
3. The one place where we were not hardwring "localhost" was when starting a peer node in tests, as we would pass only the node name, so `inet:gethostaddr/0` would be used. This would still be problematic when running tests on hosts named `nnn.my.fleet.net`, where `nnn` is a number, as the node would then be called `debuggee@nnn`, and trying to connect to such a node would fail, since `nnn` is then interpreted as an IPv4 address :shrug:

# This diff
So we fix all these issues at once, making sure (almost) no hardwired "localhost" remains. Problem 3 is the trickiest to solve, For this, we introduce a new `edb_node_monitor:safe_sname_hostnane/0` function that tries to use `inet:gethostname/0` but checks that the resulting host can be resolved back to some IP associated with the current host. If not, we fallback to localhost alternatives, always ensuring that resolution works (otherwise, we will hit hard to debug failures).

In the process, I move the calls to `net_kernel:start()` to the new API (the previous one is deprecated), and add a couple of good defaults.

Reviewed By: TheGeorge

Differential Revision: D69851260

fbshipit-source-id: 05a95a8391fc677249955ddb49cfb39601a9868d