Move distribution handling to edb:attach/1
Summary:
# Context
In order to attach to a node we need have a working distribution.

# Problem
So far, ensuring a suitable distribution exists was happening in the DAP server, but this will force us to repeat ourselves multiple times.

# This diff
Move the logic to `edb:attach/1`. Because of this, we no longer need to call ensure_distributed() when starting a peer node. The only downside is that the cookie at the first call to edb:attach() in a suite will be different from the one afterwards, so to above flakiness depending on which test runs first, we pass the cookie explicitly always

Reviewed By: TheGeorge

Differential Revision: D69870590

fbshipit-source-id: 1d0dccb90657fb0b43eeceb802e2c48627d7e8f1