Add testing option to start peer node without distribution
Summary:
# Context
So far, we start peer nodes during tests to use them as debuggee nodes, so that it is safe to pause them during tests, etc. Because we can only attach to them by dist, it makes sense that they are given a node name on creation.

# Problem
We would like to be able to write tests where debugger runs in a node without distribution, so we can check things like distribution being dynamically started, etc.

# This diff
Add a new function to `edb_test_support` that starts a peer but without distribution. While `peer` will not start distribution if a node name is not given, ?CT_PEER() will always pass a name. So the workaround is to kill distribution the hard way once the node is up.

Reviewed By: alanz

Differential Revision: D69926376

fbshipit-source-id: f1d5cc1da1bd4bb4a1cbb8a725307f8364a65275