Add a reverse-attach functionality
Summary:
# Context
At the moment, one can "attach" the edb node to a running node. This makes sense for example to debug a running program, like a webserver.

We are currently using the "attach" primitive from the DAP server to implement the "launch" request. In this type of request, the debugger is expected to start a program and leave it in a paused state.

# Problem
This is wrong. For example, when debugging a test, we start the program to test and concurrently try to set breakpoints, which is subject to race conditions. It also forces us to ensure the node we launch has a name we know in advance

# This diff
Ideally, we'd want to start a new node and leave paused and attached to the debugger node. To enable this behaviour we add a new function: `edb:reverse_attach()`.

When calling this function, we get two things:

1. A short snippet of Erlang code, that we must make sure  the new node executes (.e.g using `-eval`). On executing this code, the new node will register with the debugger ("reverse-attach") and get paused.
2. A function we can use to wait for the node to be attached, or a notification in case this times out.

We will use this function from the DAP server next.

# Implementation details

* Each time `edb:reverse_attach()` is called, a new `edb_gatekeeper` gen_server is started, under a unique name. This is responsible of:
    * Receiving the notification from the new node, via the injected code, and forwarding it to the edb_node_monitor, that is waiting to be reverse-attached
    * Resolve the call for the "wait" function returned by `edb:reverse_attach()`

* The edb_node_monitor is a state monitor that previously could be waiting to attach to a node; we generalize this so that it can be waiting for an attach or a reverse-attach

Reviewed By: TheGeorge

Differential Revision: D69991399

fbshipit-source-id: 7d39a0c1cd491d760b750240ede9c15d6a0b26aa