On attaching, validate name-domain of debugger and debuggee match
Summary:
# Context
Erlang has this (very annoying) distinction between longnames and shortnames, essentially, using fqdn or simple hostnames. Nodes with shortnames can't connect to nodes with longnames and vice-versea.

# Problem
So what happens if from a debugger node with shortnames we try to attach to a node that uses longnames? Because we keep retrying on attach errors, we'll do nothing until the timeout (if any) hits, and then just say the node was down.

As API, this can be super confusing. Instead we would like to fail early letting the user there is an issue

# This diff
We add a name-domain check and if they don't match, we raise an invalid_node error.

Reviewed By: TheGeorge

Differential Revision: D69985902

fbshipit-source-id: 7de1c6ac3c805df09363916126cd96e243c71c05