Add option to pause debuggee on bootstrap
Summary:
# Context
We use the edb_bootstrap module to install edb on the debuggee node, and start the edb_server.

# Problem
At the moment, we start the server and return, which is for when we want to attach to a running node. But we want to add support for starting a node and have it "immediately paused", which will rely on injecting a code-snippet when starting the VM. In that case, we want to pause the VM during the bootstrapping process, so we can have guarantees about when the pause happens.

# This diff
We add an option to the bootstrap option to control whether we want to pause or keep running. For now, we only use the latter

Reviewed By: TheGeorge

Differential Revision: D69991398

fbshipit-source-id: d3f58d39547d1ba695d0c34aae8edba60fc913dd