Put all launch-configuration under a common field
Summary:
# Context
When using the launch request from VS Code, the DAP server is sent the full "launch configuration". At the moment, we include some edb as top-level fields, and there is other random stuff that is sent by VS Code and we just ignore

# Problem
At some point we added strong validation for the parameters, including trying to detect typos by the user and report them as "unknown keys". However, this means that now we need to allow-list all the random stuff that VS Code sent us, which doesn't scale.

# This diff
We now make it so that we expect all "adapter configuration" all under a "config" key. So typically, launch configurations will look like:

```
{
    "name": "My config",
    "type": "erlang-edb",
    "request": "launch",
    "config": {
        // All configuration stuff expected here
    }
}
```

Because we also need to change the extensions, we will support the old format for a while as well

Reviewed By: robertoaloi

Differential Revision: D70081144

fbshipit-source-id: 7096bdb4ea6fe1b78ea06421e0e7ec428ef728da