Make attachment suites more deterministic
Summary:
# Context
In edb_attachment_SUITE we have all tests related to the debugger attaching, detaching and reattaching to a debuggee node. Currently, we spin up fresh peer nodes for the debuggee, while the debugger logic runs on the node executing the suite.

# Problem
It used to be the case `edb:attach/1` assumed the node had distribution enabled. This changed in D69870590, and now `edb:attach/1` will start distribution as a side-effect, if needed.

This affects edb_attachment_SUITE negatively in a couple of ways:
  1. The code-paths exercised by the first test now change depending on whether the node executing the suite had distribution enabled or not. In CT, the node has initially no distribution, while on buck2's test-shell, they initially do have distribution.
  2. A testcase now may have the side-effect of starting distribution. So typically the second test to run will have it enabled, but this may change if the test is ran in isolation, or if the previous test suddenly fails before attaching
  3. It's not clear how to write tests that expect distribution to be enabled/disabled

# This diff
We make it so that testcases now run on a separate node, in particular one where distribution is initially turned off. This way, we regain determinism / reproducibility

Reviewed By: TheGeorge

Differential Revision: D69982483

fbshipit-source-id: fe00d24fefc9329533e6561a5fb00b022138f2e8