Make logs easier to read
Summary:
# Context
When running the DAP server, logs including all messages are sent to a file (e.g., ~/.cache/edb/log/edb.log on Linux). These are very useful for debugging/troubleshooting.

# Problem
They are very hard to read:

  - We are logging all messages, including debug level
  - In debug level we include the raw json of every message sent/received, as a string. This is useful for debugging the framing protocol, which we are unlikely to need again, but otherwise just add noise
  - It is not easy to tell at a glance what messages are incoming and which are outgoing

# This diff
We set the default level to "info", and add "info" level logging for the Erlang representation of each incoming/outgoing message. We mark the direction succinctly with arrows like `==>` and `<==`

Reviewed By: TheGeorge, robertoaloi

Differential Revision: D70009620

fbshipit-source-id: 558d69067cd9c19f8f43e8d9b733277c3051f403