Use net_kernel:get_state() as source of truth for distribution checks
Summary:
# Context
When attaching/reverse-attaching, we need to check if distribution is started and if the domain of the debuggee matches that of the local node.

# Problem
We were using proxy values like whether `node()` is `nonode@nohost` to check the state of distribution, which is not really correct, as even with distribution started, we may not have a node name until the first connection.

A similar problem was parsing the local node name to figure out the name-domain, while we could just ask the net-kernel.

# This diff
Change the logic slightly, to use `net_kernel:get_state()` for this

Reviewed By: TheGeorge

Differential Revision: D70169170

fbshipit-source-id: a9f52192139e042d4d8a75d50438974eb44e2032