Refactor request handling
Summary:
# Context
A DEP server handles requests from a client (namely, the IDE). At the moment, for each possible type of request, we had a function in a monster motule `edb_dap_requests`.

# Problem
Whlle this was good to get started, it's no longer scaling that well:

  * `edb_dap_requests` is too large, with all kind of auxiliary functions, so makes it harder to follow
  * The types involved for the input and output live in another module, `edb_dap`, so one needs to jump back and forth between the too to understand what something is doing
  * Validating the input args is currently done in a very ad-hoc way, inside `edb_dap_server`, during dispatch
  * Similarly, we have currently no easy way to check that the client is not calling a request when we are not expecting it

# This diff
As a step towards improving this, we refactor this logic, aiming to modularize things a bit. Main changes are:

  - We introduce a `edb_dap_request` behaviour, that corresponds to a request handler. Implementations need to say how to parse their arguments and how to handle them
  - For each request `foo`, we introduce a `edb_dap_request_foo` handler that implements the above behaviour.

Essentially, we are splitting `edb_dap_requests` in multiple smaller modules, that contain types and and implementations for each request.

Here we deal with requests only, but we should reorganize reverse-requests and events in a similar way

Reviewed By: TheGeorge

Differential Revision: D70187611

fbshipit-source-id: a146814ec209f38e51440e64afda9900736651d5