Add module for validating request arguments
Summary:
# Context
At the moment, we are only validating the arguments for the "launch" request, since it is the only request where we define the schema, so we can't rely on the client being correct.

We have an ad-hoc module `edb_dap_launch_config` to handle the validation.

# Problem
We would like to be able to validate the arguements to other request. In particular, when adding support for the `attach` request, we will have the same problem of the argument not being part of the DAP spec. At the moment we can't reuse the support we have for the `launch` request

# This diff
We split the `edb_dap_launch_config` module into generic parsing utils, and move the logic that depends on the `launch` request to the respective module.

Reviewed By: TheGeorge

Differential Revision: D70259411

fbshipit-source-id: 8527950fa7d53a25b74db95159b0910c3faf286d