Simplify state handling
Summary:
# Context
It used to be the case that edb_dap_server handled both transport and logic and had a state split in two. At that point we wanted to make sure that the transport part didn't end up depending on the logic part, and the logic state was made "abstract".

Later the module grew in size as we started to store "parsed" versions of types like `target_node()` (we received `binary()` but wanted `atom()`, etc),

# Problem
We have recently solved both issues: edb_dap_transport handles transport, and we now how ways to selectively parse the json representation to get atoms when we want them etc.

So at this point, edb_dap_state offers no real value, and actually makes writing code harder, since we can't easily pattern match on it in guards, etc.

# This diff
We make server state a non-abstract type, simplifying a bit of code in the process.

NB. Now that the state is not abstract, we can have more precise types, e.g. having the context show up only on certain states. We shall do this as a follow-up

Reviewed By: TheGeorge

Differential Revision: D70331065

fbshipit-source-id: e228c8f4eaf3d6b53c00e919f883d0e2643d0328