Make state type more precise
Summary:
# Context
The DAP protocol goes through several faces: first one waits for an initialization request; then for a launch/attach request; etc. On each state, we have more information than on the previous one.

The state isrepresented by the `edb_dap_server:state()` type

# Problem
The current representation loses information: we don't know which fields are mandatory in which state, etc. This makes eqwalizer less helpful and can lead to typing issues. It is also arguably harder to understand from looking at the state how it is supposed to work

# This diff
We improve the representation of the state type so that it looks closer like a state-machine, and each state has exectaly the information that is known at that point.

Doing this actually simplifies some of the handlers,

Reviewed By: TheGeorge

Differential Revision: D70342434

fbshipit-source-id: ee4899307979d4a154351c970cf29b8d7ec744d2