Use a real parser for arguments
Summary:
# Context
At the moment we are validating the inputs to the `launch` request (even though we call it "parse").

# Problem
Because we only validate, we have fields that should be represented as atoms but come as binary from the JSON representation of the protocol. So we end up adding extra boilerplate for the conversions elsewhere

# This diff
Use a real parser for processing arguments. Down the road, this will enable to, optionally, parse other types of requests, so we don't need to do conversions later on.

Reviewed By: TheGeorge

Differential Revision: D70266744

fbshipit-source-id: 8fd6adab2474d6c792e1d80fd54671f2dca2d979