Simplify error-handling
Summary:
# Context
All handlers in the DAP server (for requests, responses and internal events), return "reaction" terms, that include: what to respond to the client (for requests only); actions to perform, like sending events to the client or terminating, and changes of internal state.

In particular, errors had to be handled by crafting special responses for the reactoin.

# Problem
Current issues:

1. The logic in the edb_dap_server itself was a bit convoluted, as we either end up with a lot of duplication, or convoluted code. In particular, having "generic error handling" means that we need to sometimes include a "response" in case of errors (for requests) or ignore them.
2. Failing in handlers is annoying, as the code needed to craft the special response is very verbose

# This diff
We make it so that "reaction" terms can contain an "error" term. This enables the following:

- The formatting of the error term can happen in only one place, each handler only needs to declaratively say "there was an error"
 - In particular, we now just "throw" in case of internal errors
  - Eqwalizer now complaints if we try to manually craft an error term inside a handler, so we will be consistent
  - Error handling in edb_dap_server becomes overall simpler

Reviewed By: TheGeorge

Differential Revision: D70250748

fbshipit-source-id: ebbd48c9c4fa3bee474b1cdcd796c2dbd06e59b5