Use an error tag for bootstrap failures during attach
Summary:
# Context
When attaching/reverse-attaching to a node, we run a bootstrap script on the debuggee node that ichecks the node is debuggable, installs the edb modules, starts edb_server, etc.

If any of these steps fails, we return a bootstrap-failure error.

# Problem
The `bootstrap_failure()` type is a sum, so for callers that want to format errors it would be convenient to identify when the error is of this type, so a proper formatting function can be called.

# This diff
We use a `bootstrap_failed` tag to wrap this type of failures

Reviewed By: TD5

Differential Revision: D70395771

fbshipit-source-id: afff3ea0ec9d64b02b62cb9f7d84f95ddf9e6626