Improve error logging
Summary:
# Context
In the DAP server, all handlers for requests, internal events, etc, return a `reaction()` term, that, in case of errors, contains an `error` key, with a structured error.

# Problem
There were several issues:

1. We were not centralizing the logging of error/warnings, so instead we were manually logging this in the handlers themselves
2. We were not consistent, so there were errors that would never get logged, and others would get logged twice

# This diff
We make it so that all error logging is done in the `edb_dap_server` itself, whenever the `reaction()` term is processed. In doing this, we had to go through al; existing handlers to remove redundant logging and notice that we were not consistent in how we handle common errors (like ids not found or processes not being suspended), so add helpers for those.

Reviewed By: TD5

Differential Revision: D70390923

fbshipit-source-id: b437340f8578ee01e195ecfa5e916f5d15eecf43