Remove "context()" from server state
Summary:
# Context
We are making the type of the DAP server's state more precise. In particular, we want to ensure that at each state, only stuff that is relevant for that is present.

# Problem
We currently have a `context()` type that is a collection of random bits. The context is available from the moment the "launch" request is received, but contains stuff that is needed only while launching, but not once we are attached..

# This diff
We take the bits of the context that are actually needed and make sure they are only available on the relevant states.

Reviewed By: TD5

Differential Revision: D70395770

fbshipit-source-id: dc46530a75c63822e1397d963c68c57e9c6a79c1